chapter on how to refactor a method part one
las two commits were on how to refactor a variable to an object, this commit looks at how to refactor the method that uses those new objects, so the code still works but basically instead of the method returning an int now it return the object, so the Testcode has to be adapted aswell not to compare the objects but rather a value to the new objectvalue, like 'assertEquals(2.00, Movie.getCharge(2).getAmount(), 0.001);'. But this is only the first five steps.