chap 5.12, refactor Price to create two instances instead of two subclasses
so this makes the two subclasses obsolete, since the changes are so minor form one to the other subclass, the book suggests to integrate them ass references in the superclass calling its own consturctor. then the changes make the other four classes obsolete, so merge the JUnit tests to Price, deleted old class and test class.